FabricaDeArmas implementada completamente.
Modificacion de tipos en Arma y derivados.
Movido el cast de flota. Aun no me convence.

+++ tp2-codigo-fuente/tp2/modelo/ArmaDirigida.java
@@ -14,7 +14,7 @@ public class ArmaDirigida extends ArmaLimitada {
 	// Inicializa el arma con la posición, escenario, velocidad de disparo, 
 	// frecuencia de disparo, dirección de disparo y la identificación recibidos. La 
 	// carga inicial es cero y no hay objetivo asignado
-	public ArmaDirigida(Point posicion, Escenario escenario, int velocidad, int frecuencia, Point direccion, String id) {
+	public ArmaDirigida(Point posicion, Escenario escenario, int velocidad, double frecuencia, Point direccion, String id) {
 		super(posicion, escenario, velocidad, frecuencia, direccion, id);
 	}
